requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
Flask==0.10.1
gunicorn==19.4.5
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.35
MarkupSafe==0.23
mongoengine==0.10.6
pymongo==3.2.1
python-dateutil==2.5.0
pytz==2016.3
requests==2.9.1
selenium==2.52.0
selenium-requests==1.2.7
six==1.10.0
tld==0.7.6
Werkzeug==0.11.5
